Make an extra module for attacks that add/remove constraints
