+++ cooked-validators/cooked-validators.cabal
@@ -14,6 +14,7 @@ library
   exposed-modules:
       Cooked
       Cooked.Attack
+      Cooked.Attack.AddConstraints
       Cooked.Attack.Common
       Cooked.Attack.DatumHijacking
       Cooked.Attack.DoubleSat
@@ -95,6 +96,7 @@ test-suite spec
       Cooked.AttackSpec
       Cooked.AttackSpec.Common
       Cooked.AttackSpec.DatumHijacking
+      Cooked.AttackSpec.DoubleSat
       Cooked.AttackSpec.DupToken
       Cooked.AttackSpec.OutPermutations
       Cooked.AttackSpec.Util
